requirements.txt
openai>=1.0.0
PyYAML>=6.0